+++ bitflyer/bitflyer_test.go
@@ -8,7 +8,7 @@ import (
 
 	exchange "github.com/openware/irix"
 	"github.com/openware/irix/config"
-	"github.com/openware/irix/faker/bitfaker"
+	"github.com/openware/irix/faker"
 	"github.com/openware/irix/portfolio/withdraw"
 	"github.com/openware/pkg/asset"
 	"github.com/openware/pkg/common"
@@ -82,7 +82,7 @@ func TestGetTransactionByHashCA(t *testing.T) {
 
 func TestGetAddressInfoCA(t *testing.T) {
 	t.Parallel()
-	v, err := b.GetAddressInfoCA(bitfaker.BitcoinTestAddress)
+	v, err := b.GetAddressInfoCA(faker.BitcoinTestAddress)
 	if err != nil {
 		t.Error("Bitflyer - GetAddressInfoCA() error:", err)
 	}
@@ -347,7 +347,7 @@ func TestCancelExchangeOrder(t *testing.T) {
 	currencyPair := currency.NewPair(currency.LTC, currency.BTC)
 	var orderCancellation = &order.Cancel{
 		ID:            "1",
-		WalletAddress: bitfaker.BitcoinTestAddress,
+		WalletAddress: faker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 		AssetType:     asset.Spot,
@@ -369,7 +369,7 @@ func TestCancelAllExchangeOrders(t *testing.T) {
 	currencyPair := currency.NewPair(currency.LTC, currency.BTC)
 	var orderCancellation = &order.Cancel{
 		ID:            "1",
-		WalletAddress: bitfaker.BitcoinTestAddress,
+		WalletAddress: faker.BitcoinTestAddress,
 		AccountID:     "1",
 		Pair:          currencyPair,
 		AssetType:     asset.Spot,
@@ -393,7 +393,7 @@ func TestWithdraw(t *testing.T) {
 		Currency:    currency.BTC,
 		Description: "WITHDRAW IT ALL",
 		Crypto: withdraw.CryptoRequest{
-			Address: bitfaker.BitcoinTestAddress,
+			Address: faker.BitcoinTestAddress,
 		},
 	}
