+++ faker/faker.go
@@ -1,3 +1,3 @@
-package bitfaker
+package faker
 
 const BitcoinTestAddress = "1MK6jowLN2kC5RrWoHMVqFi2Gq6yQjHQKm"
